Big commit
Relocate component files for organizing
Delete EventsIndex route
Add reset button to clear events
Add styling for filters
