+++ inspector/Recorder/filters.css
@@ -1,0 +1,36 @@
+section#filters {
+  display: flex;
+  flex-flow: column nowrap;
+  align-items: center;
+}
+
+#filterlabel {
+  text-decoration: none;
+  margin-bottom: 5px;
+}
+
+#filterbox {
+  display: flex;
+  flex-flow: column nowrap;
+  align-items: center;
+
+  border: 2px silver solid;
+  border-radius: 15px;
+  padding: 10px 25px;
+}
+
+#filterrow1 {
+  display: flex;
+  flex-flow: row nowrap;
+  flex: 1;
+}
+
+#filterrow2 {
+  display: flex;
+  flex-flow: row nowrap;
+  flex: 1;
+}
+
+#filter {
+  padding: 10px 15px;
+}
