+++ inspector/Recorder/recorder.css
@@ -1,0 +1,13 @@
+#recorder {
+  display: flex;
+  flex-flow: column nowrap;
+  align-items: center;
+}
+
+#resetbutton {
+  font-size: 15px;
+  padding: 10px 20px;
+  margin: 10px;
+  font-family: inherit;
+  border-radius: 10px;
+}
