+++ inspector/Recorder/recorder.tsx
@@ -1,9 +1,12 @@
 import React, { useRef } from 'react';
-import { Link } from 'react-router-dom';
+
 import { spawn, timeout } from 'effection';
 import { on } from '@effection/events';
+import { Filters } from './filters';
 
-import { useOperation, useSlice } from './hooks';
+import { useOperation, useSlice } from '../hooks';
+
+import './recorder.css'
 
 export function Recorder() {
   let [,events] = useSlice<Event[]>('events');
@@ -11,8 +14,6 @@ export function Recorder() {
   let field1 = useRef<HTMLInputElement | null>(null);
 
   useOperation(function*() {
-    events.set([]);
-
     while (!field1.current) {
       yield timeout(1);
     }
@@ -31,26 +32,33 @@ export function Recorder() {
   });
 
   return (
-    <>
-      <Link to="/">Stop</Link>
-
-      <form>
+    <div id="recorder">
+      <form id="event-fields-form">
         <p>
           Play around with these fields to see the sequences of events that the browser emits
         </p>
         <fieldset>
-          <label htmlFor="field1">Field 1</label>
+          <label htmlFor="field1">Input Text (w/ ref)</label>
           <input ref={field1} type="text" id="field1"/>
         </fieldset>
         <fieldset>
-          <label htmlFor="field2">Field 2</label>
-        <input type="text" id="field2"/>
+          <label htmlFor="field2">Input Text (w/o ref)</label>
+          <input type="text" id="field2"/>
         </fieldset>
         <fieldset>
-        <label htmlFor="field3">Field 3</label>
-        <textarea id="field3"></textarea>
+          <label htmlFor="field3">TextArea</label>
+          <textarea id="field3"></textarea>
         </fieldset>
       </form>
-    </>
+      <Filters/>
+      <button 
+        onClick={() => events.set([])} 
+        type="reset" 
+        form="event-fields-form"
+        id="resetbutton"
+      >
+        RESET
+      </button>
+    </div>
   );
 }
