--- inspector/app.css
@@ -1,33 +1,0 @@
-html {
-  height: 100%;
-}
-h1,h2,h3,h4,h5,h6 {
-  padding: 0;
-  margin: 0;
-}
-#app-element {
-  height: 100%;
-  display: flex;
-  flex-flow: row nowrap;
-}
-
-section h2 {
-  border-bottom: 1px solid black;
-  margin: 0 1em;
-}
-
-section#filters {
-  border-right: 1px solid black;
-  padding-right: 1em;
-}
-
-section#events {
-  border-right: 1px solid black;
-  width: 10em;
-}
-
-section#events Ul {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-}
