--- inspector/events-index.tsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
-
-export function EventsIndex() {
-  return (
-    <>
-      <Link to="record">Record</Link>
-    </>
-  );
-}
