+++ inspector/index.css
@@ -1,0 +1,22 @@
+html {
+  height: 100%;
+  background-color: #212324;
+  color: white;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+    sans-serif;
+}
+
+h1,h2,h3,h4,h5,h6 {
+  padding: 0;
+  margin: 0;
+}
+
+#app-element {
+  display: flex;
+  flex-flow: row nowrap;
+
+  height: 100%;
+  
+  align-items: center;
+}
