+++ inspector/index.html
@@ -1,5 +1,5 @@
 <html>
-  <link rel="stylesheet" href="./app.css"/>
+  <link rel="stylesheet" href="./index.css"/>
   <title>Key Inspector</title>
   <body>
     <div id="app-element"></div>
